updated the dio wrapper for server calls
Dio now has more informative error codes, letting someone using the API know if the errors originated from the server network or client code.

The app now directs users to the sign in page if they are connected to the server but not signed in. letting them draw if they are disconnected from the server or already signed in